Change to Maze Algorithm
No longer using a 4 state prediction technique. It will use 2 states (-1 is left, 0 is default, 1 is right). It will be changed so if the left touch sensor is pressed, it will turn right. If the right sesnor is pressed, it will turn left. If both sensors are pressed, then it will try left first, then right, then right again. The successful turn will be recorded in the history  for future iterations.