Check if both sensors are pressed function
Added a function called both_touch_press() that determines if both sensors are pressed at the same time. This function was added to resolve a problem where often times, the robot would realize one of the sensors was pressed in at the same time rather than both. To fix this, we added a short sleep timers to check if both touch sensors had been pressed.